% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2SSPmodel.R
\name{CPCM2SSP}
\alias{CPCM2SSP}
\title{Check Proportional Hazards for a given model}
\usage{
CPCM2SSP(
  time,
  event,
  group,
  changepoints,
  silent,
  xlab = "",
  ylab = "",
  main = "",
  cex.axis,
  cex.lab,
  lty,
  legend.inset,
  legend.cex,
  lwd
)
}
\arguments{
\item{time}{Time to event or censoring.}

\item{event}{An indicator vector with values of 1 for individuals who had the event occur or 0 if the participant was censored.}

\item{group}{An indicator vector with values of 1 if the the participant was in the treatment arm and 0 otherwise.}

\item{changepoints}{A vector with times where we will cut as part of a time-varying covariate.}

\item{silent}{Logical argument, FALSE indicates the user wants plots and TRUE indicates no plots only AUC calculations (default is FALSE).}

\item{xlab}{String argument for the horizontal axis label of the ROC curve.}

\item{ylab}{String argument for the vertical axis label of the ROC curve.}

\item{main}{String argument for the title of the ROC curve.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{lty}{Optional graphical parameter to set the type of line to use. Can be a number or a vector. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling the inset of the legend.}

\item{legend.cex}{Optional graphical parameter for magnification of the legend's text.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}
}
\value{
A plot of the curve and object containing:
\itemize{
 \item A cox model object based on the change points.
 \item The area under the curve for the given plot.
}
}
\description{
Check Proportional Hazards for a given model
}
