% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKMtab.R
\name{getKMtab}
\alias{getKMtab}
\title{INTERNAL: data management not for user}
\usage{
getKMtab(time, event, group)
}
\arguments{
\item{time}{Numeric or character vector of subject's unique identifier (i).}

\item{event}{Vector indicating the observation or episode (j) for a subject (i). This will determine order of events for each subject.}

\item{group}{Vector with the lengths of time spent in event of Type I for individual i in episode j.}
}
\value{
A list including:
\itemize{
 \item skm: a matrix with event times, estimated survival probs and indicator for ties.
 \item mskm: the lowest value for estimated survival.
}
}
\description{
This function gets KM survival estimators for each group.
}
\keyword{internal}
