% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intROCsurv.R
\name{intROCsurv}
\alias{intROCsurv}
\title{Create an ROC curve for survival data from clinical trials.}
\usage{
intROCsurv(
  left,
  right,
  group,
  iterations,
  end_follow,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  cex.axis = 1.5,
  cex.lab = 1.5,
  legend.inset = 0.02,
  legend.cex = 1.5,
  lty = c(1, 2, 3),
  lwd = 1.5
)
}
\arguments{
\item{left}{Time at the left of the interval (L,R]}

\item{right}{Time at the right of the interval (L,R]}

\item{group}{An indicator vector with values of 1 if the the participant was in the treatment arm and 0 otherwise.}

\item{iterations}{Number of maximum iterations for the EM algorithm. Default is 1000.}

\item{end_follow}{A numeric value indicating the end of study follow-up. Default is the largest value for \code{right}.}

\item{xlab}{String argument for the horizontal axis label of the curve.}

\item{ylab}{String argument for the vertical axis label of the curve.}

\item{main}{String argument for the title of the curve.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling legend inset.}

\item{legend.cex}{Optional graphical parameter for magnification of a legend's text.}

\item{lty}{Optional graphical parameter to set the type of lines to use. Can be a number or a vector. See \link[graphics]{par} for more details.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}
}
\value{
A plot of the ROC curve and an ROCsurv object containing:
\itemize{
 \item Data frame with the estimated survival curves.
 \item icsurv object from application of EM to the control group data.
 \item icsurv object from application of EM to the treated group data.
}
}
\description{
This function creates ROC curve for survival data from clinical trials and calculates the area under the curve.
}
