% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2SSPrecur.R
\name{rec2Ssurv}
\alias{rec2Ssurv}
\title{Create an ROC curve for recurrent event data that meets exchangeability assumption for a clinical trials.}
\usage{
rec2Ssurv(
  id,
  episode_j,
  time,
  group,
  mi,
  checkPH = FALSE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  cex.axis = 1.5,
  cex.lab = 1.5,
  legend.inset = 0.02,
  legend.cex = 1.5,
  lty = c(2, 1, 3),
  lwd = 1.5
)
}
\arguments{
\item{id}{Unique identifier for each participant.}

\item{episode_j}{Vector identifying the event number (j=1,..., m_i) for each partipant.}

\item{time}{Gap time (Y_ij)}

\item{group}{An indicator vector with values of 1 if the the participant was in the treatment arm and 0 otherwise.}

\item{mi}{Vector with number of episodes/recurrences each participant experienced.}

\item{checkPH}{Logical argument to indicate if user wants to compare the nonparametric two sample survival curve with a proportional hazards model based curve (default is FALSE).}

\item{xlab}{String argument for the horizontal axis label of the ROC curve.}

\item{ylab}{String argument for the vertical axis label of the ROC curve.}

\item{main}{String argument for the title of the ROC curve.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling the inset of the legend.}

\item{legend.cex}{Optional graphical parameter for magnification of the legend's text.}

\item{lty}{Optional graphical parameter to set the type of line to use. Can be a number or a vector. See \link[graphics]{par} for more details.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}
}
\value{
A plot of the ROC curve (if \code{silent=FALSE}) and an ROCsurv object containing:
\itemize{
 \item A survfit object containing the survival curve for the treatment group.
 \item A survfit object containing the survival curve for the control group.
 \item The area under the curve for the ROC in the given plot.
}
}
\description{
This function creates ROC curve for survival data from clinical trials and calculates the area under the curve.
}
\details{
The methods avaiable are "restrict" or "complete"
}
