% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCmodification.R
\name{ROCmodifier}
\alias{ROCmodifier}
\title{ROC when survival goes to 0 for either group}
\usage{
ROCmodifier(
  time,
  event,
  group,
  modifier,
  area = FALSE,
  mlabels,
  xlab,
  ylab,
  main
)
}
\arguments{
\item{time}{Numeric vector with times to event.}

\item{event}{Vector of indicators with value 1 subject experienced the event ocurred and 0 when censored.}

\item{group}{Vector of indicators for the treatment group with value 1 for treatment and 0 for control.}

\item{modifier}{Vector of indicators for the modifier with values of 1 when modifier is present and 0 otherwise.}

\item{area}{TRUE/FALSE argument to indicate if user wants an estimate for area under the curve.}

\item{mlabels}{A vector with two strings to label the levels (0, 1) of the modifier.}

\item{xlab}{passed from ROCsurv}

\item{ylab}{passed from ROCsurv}

\item{main}{passed from ROCsurv}
}
\value{
A plot of the ROC curve and an ROCsurv object containing:
\itemize{
 \item A list with two survfit objects for the treatment and control groups when the modifier=0.
 \item A list with two survfit objects for the treatment and control groups when the modifier=1.
 \item If area = TRUE, the areas under the curve for the two ROCs in the given plot.
}
}
\description{
ROC when survival goes to 0 for either group
}
