% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCparametric.R
\name{ROCparametric}
\alias{ROCparametric}
\title{ROC when survival goes to 0 for either group}
\usage{
ROCparametric(time, event, group, dist = "lognormal", silent, xlab, ylab, main)
}
\arguments{
\item{time}{Numeric or character vector of subject's unique identifier (i).}

\item{event}{Vector indicating the observation or episode (j) for a subject (i). This will determine order of events for each subject.}

\item{group}{Vector with the lengths of time spent in event of Type I for individual i in episode j.}

\item{dist}{String indicating the distribution.}

\item{xlab}{passed from ROCsurv}

\item{ylab}{passed from ROCsurv}

\item{main}{passed from ROCsurv}
}
\value{
A plot of the ROC curve and an ROCsurv object containing:
\itemize{
 \item A survfit object for the treatment group.
 \item A survfit object for the control group.
 \item The area under the curve for the ROC in the given plot.
}
}
\description{
ROC when survival goes to 0 for either group
}
