% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSurvTrial.R
\name{simSurvTrial}
\alias{simSurvTrial}
\title{Simulate a clinical trial with survival outcomes for treatment and control groups based on various proportional hazards models.}
\usage{
simSurvTrial(
  size,
  followup,
  beta,
  X = NULL,
  dist = NULL,
  params = NULL,
  hazard = NULL,
  censprop = NULL
)
}
\arguments{
\item{size}{Numeric value specifying the number of subjects or sample size.}

\item{followup}{Numeric value specifying the simulated study's follow-up time in years.}

\item{beta}{Numeric value fot the treatment effect.}

\item{X}{Optional data frame with two columns for subject id's and their assigned treatment.
The column for treatment assignment must be a logical or numeric vector where values of TRUE or 1 will indicate individual was assigned to the treatment
arm and FALSE or 0 to control arms. Any other values are invalid.
If $X$ is not provided the treatment will be assigned randomly with probability 0.5 and subject id's will be a sequence from 1 to the specified sample size.}

\item{dist}{Optional character string specifying the parametric survival distribution to be used. Options include "weibull" (the default), "exponential", or "gompertz".
This is ignored if the \code{hazard} argument is provided. See \strong{Details} and \strong{Examples}.}

\item{params}{Optional numeric value or vector with paramaters for the chosen distribution where the first paramater will correspond to the shape \eqn{\lambda} for the
and the second to the scale \eqn{\gamma} of the Weibull, Exponential (only uses the first), and Gompertz distributions.}

\item{hazard}{Optional user-defined hazard function, with arguments \code{t}, \code{X}, and \code{beta}. This function should return
the hazard at time \code{t} for an individual with treatment assigment supplied via \code{X} and treatment effect supplied via \code{beta}.
See \strong{Details} and \strong{Examples}.}

\item{censprop}{Optional the proportion of observations to designate as right-censored when the flexible-hazards model is used. This is ignored if the \code{hazard}
or \code{dist} arguments are provided. See \strong{Details} and \strong{Examples}.}
}
\value{
A data frame containing:
\itemize{
 \item subject id's
 \item Treatment assignment 0-control, 1-treatment
 \item Time to event (time)
 \item Event indicator (event: 1-event occurred, 0-no event/censored).
}
}
\description{
This function uses the simsurv and coxed packages to simulate event and censoring times for a proportional hazards
model with either: dparametric distribution assumptions, or a user defined hazard function, or in using a flexible-hazard
method that needs no distributions to be specified. The user can provide treatment assignments for the subjects or have
them randomly assigned to each subject.
}
\examples{

# Simulate a clinical trial with 500 subjects that were followed for 2 years.
# where the true hazard ratio of 0.2 indicates a large treatment effect.
n = 500
maxt = 2
beta = log(0.2)

#1) Assuming data follows Wiebull(1, 1.5).
    set.seed(28)
    simdata <- simSurvTrial(size = n, followup = maxt, beta = beta, dist = "weibull",
    params = c(1, 1.5))
    # Check the maximum survival time in the data the censoring rate.
     maxt <- max(simdata$time)
     maxt
     cens <- table(simdata$event)[1]/500 #estimate proportion censored
     cens

#2) Assuming no distribution and using flexible hazards model with 25\% censoring rate.
    set.seed(28)
    simdata2 <- simSurvTrial(size = n, followup = maxt, beta = beta, censprop = 0.25)
    # Check the maximum survival time in the data the censoring rate.
     maxt <- max(simdata2$time)
     maxt
     cens <- table(simdata2$event)[1]/500 #estimate proportion censored
     cens

}
