% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate a clinical trial with survival outcomes for treatment and control groups based on proportional hazards.}
\usage{
simulate(size, time, hazardratio, censprop)
}
\arguments{
\item{size}{The total number of subjects or sample size.}

\item{time}{The simulated study's follow-up time (maximum time a subject could survive without censoring).}

\item{hazardratio}{The treatment to control hazard ratio.}

\item{censprop}{Proportion of observations to designate as right-censored.}
}
\value{
A data frame containing:
\itemize{
 \item Treatment assignment 0-control, 1-treatment
 \item Time to event (time)
 \item Event indicator (event: 1-event occurred, 0-no event/censored).
}
}
\description{
This function randomly assigns individuals to treatment or control groups then uses
the coxed package to simulate event and censoring times based on a proportional hazards
model without any other parametric asumptions based on the flexible-hazard method.
See Harden and Kropko (2018) for further details.
}
\examples{

#Simulate a study of 500 subjects followed for 365 days where the hazard ratio of 0.2 shows treatment prolongs life.
set.seed(28)
n=500
simdata <- simulate(n, 365, 0.2, 0.15)

#Check proportion censored and hazard ratio estimates and compare to true.
cens <- unname(table(simdata$event)[1]/n) #estimate proportion censored
treatment_hr <- unname(exp(coxph(Surv(time, event) ~ treatment, data=simdata, ties="breslow")$coefficients))
cens
treatment_hr

}
