% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWOSCI.R
\name{TwoSampCI}
\alias{TwoSampCI}
\title{Create a two-sample cumulative incidence curve for data from randomized controlled trials.}
\usage{
TwoSampCI(
  time,
  event,
  group,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  rlabels,
  cex.axis = 1.5,
  cex.lab = 1.5,
  lwd = 1.5,
  legend.inset = 0.02,
  legend.cex = 1.5
)
}
\arguments{
\item{time}{Time to an event of any kind or to censoring.}

\item{event}{An indicator vector with values of 1, 2, ..., M related to one of M possible causes for an event for an participant that had an event occur or 0 if the participant was censored.}

\item{group}{An indicator vector with values of 1 if the the participant was in the treatment arm and 0 otherwise.}

\item{xlab}{String argument for the horizontal axis label of the ROC curve.}

\item{ylab}{String argument for the vertical axis label of the ROC curve.}

\item{main}{String argument for the title of the ROC curve.}

\item{rlabels}{A vector with M strings to label the curves related to each of the M causes for events.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling the inset of the legend.}

\item{legend.cex}{Optional graphical parameter for magnification of the legend's text.}
}
\value{
A plot of the ROC curve (if \code{silent=FALSE}) and an ROCsurv object containing:
\itemize{
 \item A Cuminc class object containing survival and cummulative incidence estimates and corresponding standard errors for each group.
 \item A matrix representation of the two-sample cumulative incidence curve \code{C_u}.
}
}
\description{
This function creates a two-sample cumulative incidence curve for data from randomized controlled trials and has an option to diagnose the proportional hazard assumption.
}
\details{
The methods avaiable are "restrict" or "complete"
}
