% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2SCI.R
\name{TwoSCI}
\alias{TwoSCI}
\title{Create a two-sample cumulative incidence curve for data from randomized controlled trials.}
\usage{
TwoSCI(
  time,
  event,
  group,
  maxt = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  rlabels,
  cex.axis = 1.5,
  cex.lab = 1.5,
  lwd = 1.5,
  silent = FALSE,
  lty = c(2, 3, 1),
  legend.inset = 0.02,
  legend.cex = 1.5,
  checkFG = FALSE,
  c_index = TRUE,
  CI = FALSE,
  level = 0.95,
  B
)
}
\arguments{
\item{time}{Time to an event of any kind or to censoring.}

\item{event}{An indicator vector with values of 0, 1, 2, ..., M indicating one of M possible causes for an event for an participant that had an event occur or 0 if the participant was censored.}

\item{group}{An indicator vector with values of 1 if the the participant was in the treatment arm and 0 otherwise.}

\item{maxt}{Duration of the trial.}

\item{xlab}{String argument for the horizontal axis label of the ROC curve.}

\item{ylab}{String argument for the vertical axis label of the ROC curve.}

\item{main}{String argument for the title of the ROC curve.}

\item{rlabels}{A vector with M strings to label the curves related to each of the M causes for events.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}

\item{silent}{Logical argument, FALSE indicates the user wants plots and TRUE indicates no plots only calculations (default is FALSE).}

\item{lty}{Optional graphical parameter to set the type of line to use. Can be a number or a vector. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling the inset of the legend.}

\item{legend.cex}{Optional graphical parameter for magnification of the legend's text.}

\item{checkFG}{Logical argument to indicate if the user wants to compare the nonparametric curve with the curve based on the Fine-Gray model (default is FALSE).}

\item{c_index}{Logical argument to indicate if the user wants an estimate for the concordance for each risk (default is TRUE).}

\item{CI}{Optional logical argument to indicate if the user wants a bootstrap confidence interval for the curves and the concordance.}

\item{level}{A numerical argument to indicate the confidence level for the confidence interval. Default = 0.95.}

\item{B}{The number of bootstrap samples to use for the confidence interval. Default = 1000.}
}
\value{
A plot of the curve (if \code{silent=FALSE}) and an object containing:
\itemize{
 \item A Cuminc object containing survival and cummulative incidence estimates and corresponding standard errors for each group.
 \item List with the two-sample cumulative incidence curves \code{C_u} for each risk type.
}
}
\description{
This function creates a two-sample cumulative incidence curve for data from randomized controlled trials and has an option to diagnose the proportional hazard assumption.
}
