% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{TwoSPC-package}
\alias{TwoSPC-package}
\alias{_PACKAGE}
\title{A  Package  for  the  Analysis  of  Time  to  Event  OutcomesSubject  to    Right-Censoring  Using  ROC  Curves  in  Two-Arm  Randomized  Trials.}
\description{
Survival data is often used in clinical trials to determine effectiveness of a new treatment.
Analysis of such data is usually completed using two separate Kaplan-Meier curves, a Log Rank Test and Cox Proportional Hazard models.
We propose an alternative way to analyze this type of data by combining the two Kaplan-Meier curves into an ROC curve and using the area under the curve (AUC) as measure of treatment effectiveness.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SandraCastroPearson/2SPcurves}
  \item Report bugs at \url{https://github.com/SandraCastroPearson/2SPcurves/issues}
}

}
\author{
Sandra Castro-Pearson, Xianghua Luo, Chap Le
}
