% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2SSPdcens.R
\name{TwoSSPdcens}
\alias{TwoSSPdcens}
\title{Create an ROC curve for survival data from clinical trials.}
\usage{
TwoSSPdcens(
  o_l,
  o_r,
  e_l,
  e_r,
  group,
  maxtime = NULL,
  ctime = 99,
  dtype = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  cex.axis = 1.5,
  cex.lab = 1.5,
  legend.inset = 0.02,
  legend.cex = 1.5,
  lty = c(1, 2, 3),
  lwd = 1.5
)
}
\arguments{
\item{o_l}{Time at the left of the interval (L,R] for the onset time (see details).}

\item{o_r}{Time at the right of the interval (L,R] for the onset time (see details).}

\item{e_l}{Time at the left of the interval (L,R] for the event time (see details).}

\item{e_r}{Time at the right of the interval (L,R] for the event time (see details).}

\item{group}{Indicator vector with values of 1 if the the participant was in the treatment arm and 0 otherwise.}

\item{maxtime}{Maximum event time in the sample.}

\item{ctime}{Last time an event could have been observed in the study (length of the study, must be higher than maxtime even if by 1 unit).}

\item{dtype}{String argument that can take values "dcens" or "DIcens" to indicate either doubly censored or doubly interval censored data (see details).}

\item{xlab}{String argument for the horizontal axis label of the curve.}

\item{ylab}{String argument for the vertical axis label of the curve.}

\item{main}{String argument for the title of the curve.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling legend inset.}

\item{legend.cex}{Optional graphical parameter for magnification of a legend's text.}

\item{lty}{Optional graphical parameter to set the type of lines to use. Can be a number or a vector. See \link[graphics]{par} for more details.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}
}
\value{
A plot of the ROC curve and an ROCsurv object containing:
\itemize{
 \item R_u: Data frame with the estimated two-sample survival curves.
 \item npmle0: NPMLE's (outcome and gap) for the untreated/unexposed group.
 \item npmle1: NPMLE's (outcome and gap) for the treated/exposed group.
}
}
\description{
This function creates ROC curve for survival data from clinical trials and calculates the area under the curve.
}
\details{
For doubly censored data as discussed by Turnbull (dtype = "dcens"):
  Set o_l = NA and o_r = NA
  Set e_l = e_r if the event is exact
  Set e_l = -Inf if the event time is left censored and e_r = Inf if it is right censored

For doubly interval censored data (dtype = "DIcens) as discussed by De Gruttola and Legakos:
  Set o_l = o_r if the onset time is exact, o_l = -Inf if it is left censored, and o_r = Inf if it is right censored.
  Set e_l = e_r if the event time is exact, e_l = -Inf if it is left censored, and e_r = Inf if it is right censored.
}
