% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2SSPipw.R
\name{TwoSSPiptw}
\alias{TwoSSPiptw}
\title{Create an inverse probability of treatment weight adjusted two-sample survival probability curve for data from observational studies.}
\usage{
TwoSSPiptw(
  time,
  event,
  group,
  weights,
  area = FALSE,
  CI = FALSE,
  level = 0.95,
  B = 1000,
  checkPH = FALSE,
  abtwc = FALSE,
  silent = FALSE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  cex.axis = 1.5,
  cex.lab = 1.5,
  legend.inset = 0.02,
  legend.cex = 1.5,
  lty = c(2, 1, 3),
  lwd = 1.5
)
}
\arguments{
\item{time}{Time to event or censoring.}

\item{event}{An indicator vector with values of 1 for individuals who had the event occurred or 0 if the participant was censored.}

\item{group}{An indicator vector with values of 1 if the the participant was in the exposed group and 0 otherwise.}

\item{weights}{A vector with weights for each study participant.}

\item{area}{Logical argument to indicate if user wants an estimate for area under the curve (default is TRUE).}

\item{CI}{Logical argument indicating if user wants bootstrap errors and confidence intervals to be calculated for the curve and AUC. Default is FALSE.}

\item{level}{Numerical argument indicating condifence level for the confidence interval. Default is 0.95. It's ignored if CI=FALSE.}

\item{B}{Number of bootstrap samples to use for confidence interval and error calculations. Default is 1000. It's ignored if CI=FALSE.}

\item{checkPH}{Logical argument to indicate if user wants to compare the nonparametric curve with the curve based on the proportional hazards model (default is FALSE).}

\item{abtwc}{Logical argument to indicate if area between curves is needed as part of model comparisons (checkPHM or compare must be TRUE).}

\item{silent}{Logical argument, FALSE indicates the user wants ROC plots and TRUE indicates no plots only AUC calculations (default is FALSE).}

\item{xlab}{String argument for the horizontal axis label of the ROC curve.}

\item{ylab}{String argument for the vertical axis label of the ROC curve.}

\item{main}{String argument for the title of the ROC curve.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling the inset of the legend.}

\item{legend.cex}{Optional graphical parameter for magnification of the legend's text.}

\item{lty}{Optional graphical parameter to set the type of line to use. Can be a number or a vector. See \link[graphics]{par} for more details.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}
}
\value{
A plot of the ROC curve (if \code{silent=FALSE}) and an ROCsurv object containing:
\itemize{
 \item A survfit object containing the survival curve for the treatment group.
 \item A survfit object containing the survival curve for the control group.
 \item The area under the curve (AUC).
 \item A matrix representation of the two-sample survival probability curve \code{R_u}.
}
}
\description{
This function creates an inverse probability of treatment weight adjusted two-samle survival probability curve for data from observational studies and has options diagnose the proportional hazard assumption.
}
