% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2SSPtrunc.R
\name{TwoSSPtrunc}
\alias{TwoSSPtrunc}
\title{Create an ROC curve for left truncated right censored data.}
\usage{
TwoSSPtrunc(
  entry,
  exit,
  event,
  group,
  area = FALSE,
  silent = FALSE,
  CI = FALSE,
  level = 0.95,
  B = 1000,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  cex.axis = 1.5,
  cex.lab = 1.5,
  legend.inset = 0.02,
  legend.cex = 1.5,
  lty = c(2, 1, 3),
  lwd = 1.5
)
}
\arguments{
\item{entry}{The entry time for the participant that determined left truncation.}

\item{exit}{Time when the event or censoring occurred (must be higher than entry time).}

\item{event}{An indicator vector with values of 1 for individuals who had the event occur or 0 if the participant was censored.}

\item{group}{An indicator vector with values of 1 if the the participant was in the treatment or exposed group and 0 otherwise.}

\item{area}{Logical argument to indicate if user wants an estimate for area under the curve (default is TRUE).}

\item{silent}{Logical argument, FALSE indicates the user wants ROC plots and TRUE indicates no plots only AUC calculations (default is FALSE).}

\item{CI}{Logical argument indicating if user wants bootstrap errors and confidence intervals to be calculated for the curve and AUC. Default is FALSE.}

\item{level}{Numerical argument indicating condifence level for the confidence interval. Default is 0.95. It's ignored if CI=FALSE.}

\item{B}{Number of bootstrap samples to use for confidence interval and error calculations. Default is 1000. It's ignored if CI=FALSE.}

\item{xlab}{String argument for the horizontal axis label of the curve.}

\item{ylab}{String argument for the vertical axis label of the curve.}

\item{main}{String argument for the title of the curve.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling legend inset.}

\item{legend.cex}{Optional graphical parameter for magnification of a legend's text.}

\item{lty}{Optional graphical parameter to set the type of lines to use. Can be a number or a vector. See \link[graphics]{par} for more details.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}
}
\value{
A plot of the ROC curve and an ROCsurv object containing:
\itemize{
 \item Data frame with the estimated survival curves.
 \item icsurv object from application of EM to the control group data.
 \item icsurv object from application of EM to the treated group data.
}
}
\description{
This function creates ROC curve for survival data from clinical trials and calculates the area under the curve.
}
