% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{ROCSSnonparam}
\alias{ROCSSnonparam}
\title{Nonparametric sample size or power based on the Area Under the ROC Curve.}
\usage{
ROCSSnonparam(
  AUC,
  HR,
  ctrl.event.rate,
  trt.event.rate,
  cens.rate = 0,
  sig.level = 0.05,
  power = 0.8,
  alternative = "two.sided",
  method = "survival"
)
}
\arguments{
\item{AUC}{Numeric value for the size of the AUC = P(T1 >  T0) that is expected. Must be between 0.55 and 1.0.}

\item{HR}{Only needed is AUC cannot be provided. Numeric value for the size of the hazard ratio.}

\item{sig.level}{Numeric value for the type 1 Error Rate. Default is 0.05.}

\item{power}{Numeric value for the target power. Must be between 0.2 and 1. Default is 0.8.}

\item{alternative}{Character string. Options: "two.sided" or "one.sided".}

\item{method}{Character string for method to be used to obtain sample size or power. Options: "parametric binormal", "nonparametric"}

\item{type}{Character string. Options: "biomarker" or "survival".}

\item{variance}{Required only when method = "parametric binormal". Numeric value for the variance to use when calculating sample size and power.}
}
\value{
The sample size for each of the two trial arms (n) or power.
}
\description{
This function finds the needed sample size to .

AUC, sig.level, power, variance, alternative = "two.sided"
}
\details{
Methods.
}
