% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCsurv.R
\name{ROCsurv}
\alias{ROCsurv}
\title{Create an ROC curve for survival data from clinical trials.}
\usage{
ROCsurv(
  time,
  event,
  group,
  method,
  checkPH = FALSE,
  compare = FALSE,
  area = TRUE,
  silent = FALSE,
  abtwc = FALSE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  KM,
  cex.axis = 1.5,
  cex.lab = 1.5,
  legend.inset = 0.02,
  legend.cex = 1.5,
  lty = c(2, 1, 3),
  lwd = 1.5
)
}
\arguments{
\item{time}{Time to event or censoring.}

\item{event}{An indicator vector with values of 1 for individuals who had the event occurred or 0 if the participant was censored.}

\item{group}{An indicator vector with values of 1 if the the participant was in the treatment arm and 0 otherwise.}

\item{method}{Method to be used to obtain a complete curve.}

\item{checkPH}{Logical argument to indicate if user wants to compare an ROC from the Cox Proportional Hazards Model to the KMROC (default is FALSE).}

\item{compare}{Logical argument to indicate if user wants want to compare the KMROC to ROM curves from a Cox model, a loglogistic AFT model and a lognormal AFT model (default is FALSE).}

\item{area}{Logical argument to indicate if user wants an estimate for area under the curve (default is TRUE).}

\item{silent}{Logical argument, FALSE indicates the user wants ROC plots and TRUE indicates no plots only AUC calculations (default is FALSE).}

\item{abtwc}{Logical argument to indicate if area between curves is needed as part of model comparisons (checkPHM or compare must be TRUE).}

\item{xlab}{String argument for the horizontal axis label of the ROC curve.}

\item{ylab}{String argument for the vertical axis label of the ROC curve.}

\item{main}{String argument for the title of the ROC curve.}

\item{KM}{Optional logical parameter. If TRUE the result will include survfit object containing survival information for each group.}

\item{cex.axis}{Optional graphical parameter for magnification of axis annotation. See \link[graphics]{par} for more details.}

\item{cex.lab}{Optional graphical parameter for magnification of x and y labels. See \link[graphics]{par} for more details.}

\item{legend.inset}{Optional graphical parameter controling the inset of the legend.}

\item{legend.cex}{Optional graphical parameter for magnification of the legend's text.}

\item{lty}{Optional graphical parameter to set the type of line to use. Can be a number or a vector. See \link[graphics]{par} for more details.}

\item{lwd}{Optional graphical parameter for line width relative to the default. See \link[graphics]{par} for more details.}
}
\value{
A plot of the ROC curve (if \code{silent=FALSE}) and an ROCsurv object containing:
\itemize{
 \item A survfit object containing the survival curve for the treatment group.
 \item A survfit object containing the survival curve for the control group.
 \item The area under the curve for the ROC in the given plot.
}
}
\description{
This function creates ROC curve for survival data from clinical trials and calculates the area under the curve.
}
\details{
The methods avaiable are "restrict" or "complete"
}
