% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCsurv.R
\name{ROCsurv}
\alias{ROCsurv}
\title{Create an ROC curve for survival data from clinical trials.}
\usage{
ROCsurv(
  time,
  event,
  group,
  level = NULL,
  method = NULL,
  checkPH = FALSE,
  compare = FALSE,
  area = NULL,
  silent,
  abtwc
)
}
\arguments{
\item{time}{Time to event or censoring.}

\item{event}{Vector indicating if the event occurred (event=1) or if the time was censored (event=0).}

\item{group}{Vector indicating if the individual was in the treatment arm (group=1) or the control arm (group=0).}

\item{level}{The confidence level for the confidence interval of the area under the curve.
Must be between 0.50 and 0.99. Default is 0.95. See details.}

\item{method}{Method to be used to obtain a complete curve.}

\item{checkPH}{to check}

\item{compare}{to compare fits}

\item{area}{logical argument to indicate if user wants an estimate for area under the curve (default is TRUE).}

\item{silent}{logical argument to indicate if user wants ROC plots (default is TRUE).}
}
\value{
A plot with the ROC curve and an ROCsurv object containing:
\itemize{
 \item A survfit object containing the survival curve for the treatment group.
 \item A survfit object containing the survival curve for the control group.
 \item The area under the curve for the ROC in the given plot.
}
}
\description{
This function creates ROC curve for survival data from clinical trials and calculates the area under the curve.
}
\details{
The methods avaiable are "restrict" or "complete"
}
