% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCsurv.R
\name{ROCsurv}
\alias{ROCsurv}
\title{Create an ROC curve for survival data from clinical trials.}
\usage{
ROCsurv(time, event, group)
}
\arguments{
\item{time}{Numeric or character vector of subject's unique identifier (i).}

\item{event}{Vector indicating the observation or episode (j) for a subject (i). This will determine order of events for each subject.}

\item{group}{Vector with the lengths of time spent in event of Type I for individual i in episode j.}
}
\value{
A plot with the ROC curve and an ROCsurv object containing:
\itemize{
 \item A survfit object containing the survival curve for the treatment group.
 \item A survfit object containing the survival curve for the control group.
 \item The area under the curve for the ROC in the given plot.
}
}
\description{
This function creates ROC curve for survival data from clinical trials and calculates the area under the curve.
}
\examples{
#Simulate a study: 500 subjects, 365 days follow-up, 0.2 hazard ratio where treatment prolongs life.
set.seed(28)
simdata <- simulate(n, 365, 0.2, 0.15)
#ROC curve
sim_roc <- with(simdata, ROCsurv(time, event, treatment))

}
